Move and rename get-tiered-products
